chunked: move cache to separate file
